DrSigil now writes directly into shared memory buffers
This change improves performance and reduces memory bandwidth contention.
Access to the shared memory is handled by a ticket lock that maintains
the order of threads waiting, in an attempt to reduce skew between
thread events. A custom futex implementation had to be used because
DynamoRIO does not provide a conditional variable implementation,
and the use of the pthread library is not allowed in DR clients.

https://github.com/DynamoRIO/dynamorio/issues/279